moved method info parsing to a separate file
